Makefile

+++ prepare_dev_env.sh
@@ -1,26 +1,30 @@
 #!/bin/bash
 
-read -p "Source code location: " source_code_location
-read -p "Enter domain name: " domain_name
-read -p "Enter internal IP: " internal_ip
-read -p "Enter internal IP mask: " internal_mask
-read -p "Enter external IP: " external_ip
-read -p "Enter external IP mask: " external_mask
-read -s -p "Enter MySQL root password: " mysql_root_password
+read -p "Source code location: "                    source_code_location
+read -p "Enter domain name: "                       domain_name
+read -p "Enter internal IP (eg. 192.168.1.1): "     internal_ip
+read -p "Enter internal IP mask (eg. 24): "         internal_mask
+read -p "Enter external IP (eg. 192.168.1.1): "     external_ip
+read -p "Enter external IP mask (eg. 24): "         external_mask
+read -p "Enter database host (eg. 192.168.1.1): "   db_host
+read -p "Enter AMQP host (eg. 192.168.1.1): "       amqp_host
+
+read -s -p "Enter MySQL root password: "            mysql_root_password
 echo ""
-read -s -p "Enter MongoDB root password: " mongo_root_password
+read -s -p "Enter MongoDB root password: "          mongo_root_password
 echo ""
-read -s -p "Enter RabbitMQ password: " rabbitmq_password
+read -s -p "Enter RabbitMQ password: "              rabbitmq_password
 echo ""
-read -s -p "Enter database user password: " db_user_pass
+read -s -p "Enter database user password: "         db_user_pass
 echo ""
-read -p "Enter database host: " db_host
-read -p "Enter AMQP host: " amqp_host
-read -s -p "Enter AMQP password: " amqp_pass
+read -s -p "Enter AMQP password: "                  amqp_pass
 echo ""
 
-current_path=$(pwd)
+
 echo "Creating Docker .env file..."
+
+current_path=$(pwd)
+
 cat << EOF > docker/.env
 ENV_FILE=.env
 
@@ -39,6 +43,7 @@ EXTERNAL_MASK=${external_mask}
 
 MYSQL_ROOT_PASSWORD=${mysql_root_password}
 
+MONGO_INITDB_USER=admin
 MONGO_INITDB_DATABASE=admin
 MONGO_INITDB_ROOT_USERNAME=root
 MONGO_INITDB_ROOT_PASSWORD=${mongo_root_password}
@@ -48,7 +53,7 @@ RABBITMQ_USER=root
 RABBITMQ_PASSWORD=${rabbitmq_password}
 
 DB_USER=root
-DB_PASS=${db_user_pass}
+DB_PASS=${mysql_root_password}
 DB_HOST=${db_host}
 DB_PORT=3306
 
@@ -56,45 +61,96 @@ AMQP_HOST=${amqp_host}
 AMQP_USER=root
 AMQP_PASS=${amqp_pass}
 EOF
+
 echo "Creating Docker .env file complete!"
 
+
 echo "Creating MongoDB keyFile..."
-chmod +x mongodb/init.sh
+
+if [[ ! -d "mongodb/data" ]];   then mkdir -p mongodb/data; fi
+if [[ ! -d "mognodb/backup" ]]; then mkdir -p mongodb/backup; fi
+
 openssl rand -base64 756 > mongodb/data/keyFile
-mkdir mongodb/backup
+
 sudo chown $(whoami):$(whoami) mongodb/init.sh
 sudo chown -R 1001:1000 mongodb/data
 sudo chown -R 1001:1000 mongodb/backup
 sudo chmod 400 mongodb/data/keyFile
+
 echo "Creating MongoDB keyFile complete!"
 
-mkdir -p mariadb/data
-mkdir mariadb/backup
-mkdir mariadb/schemas
-docker rm -f mariadb
-docker-compose -f docker/docker-compose.yml up -d mariadb
 
-docker rm -f mongodb
-docker-compose -f docker/docker-compose.yml up -d mongodb
-sleep 5
+echo "Initialising databases..."
+
+cat << EOF > mongodb/init/create_temp_admin.js
+db.createUser({ 
+    user: 'admin',
+    pwd: '${mongo_root_password}',
+    roles: [{role: 'root', db: 'admin'}]
+});
+EOF
+
+cat << EOF > mongodb/init/create_root.js
+db.createUser({
+    user: "root",
+    pwd: "${mongo_root_password}",
+    roles: [{role: "remote_role", db: "admin"}],
+});
+EOF
+
+cat << EOF > mongodb/init/create_root.js
+rs.initiate({
+    _id: "mongoreplicaset1",
+    version: 1,
+    members: [{ _id: 0, host : "mongo-mongo1.${domain_name}:27017"}]
+});
+EOF
+
+
+if [[ ! -d "mariadb/data" ]];  then mkdir -p mariadb/data; fi
+if [[ ! -d "mariadb/backup"]]; then mkdir -p mariadb/backup; fi
+if [[ ! -d "mariadb/migration" ]]; then mkdir -p mariadb/migration; fi
+
+if [[ "$(docker ps -aqf name=mariadb)" ]]; then 
+    docker rm -f mariadb
+    docker-compose -f docker/docker-compose.yml up -d mariadb
+    sleep 5
+fi
+
+if [[ "$(docker ps -aqf name=mongodb)" ]]; then 
+    docker rm -f mongodb
+    docker-compose -f docker/docker-compose.yml up -d mongodb
+    sleep 5
+fi
 
 cd mongodb
+chmod +x init.sh
 bash init.sh
 cd ..
 
+echo "Initialising databases complete!"
+
+
 echo "Init RabbitMQ..."
-docker rm -f rabbitmq
-docker-compose -f docker/docker-compose.yml up -d rabbitmq
-sleep 5
+
+if [[ "$(docker ps -aqf name=rabbitmq)" ]]; then 
+    docker rm -f rabbitmq
+    docker-compose -f docker/docker-compose.yml up -d rabbitmq
+    sleep 5
+fi
+
 cd rabbitmq
 chmod +x init.sh
 bash init.sh
 cd ..
+
 echo "Init RabbitMQ complete!"
+
 
 echo "Creating simple reverse proxy site..."
 
-mkdir nginx/sites
+if [[ ! -d "nginx/sites" ]]; then mkdir -p nginx/sites; fi
+
 cat << EOF > nginx/sites/${domain_name}.conf
 upstream 11b9509-6783478-bb37b70-f2617d0 {
         server ${internal_ip}:8989;
@@ -115,20 +171,29 @@ server {
 }
 EOF
 
-echo "Creating simple reverse proxy site complete!"
 echo "Your app should be listening on port 8989"
+echo "Creating simple reverse proxy site complete!"
+
 
 echo "Creating self signed SSL certificate..."
-mkdir nginx/ssl
+
+if [[ ! -d "nginx/ssl" ]]; then mkdir -p nginx/ssl; fi
 openssl req -x509 -nodes -days 3650 -newkey rsa:2048 -keyout nginx/ssl/${domain_name}.key -out nginx/ssl/${domain_name}.crt
+
+if [[ "$(docker ps -aqf name=reverse_proxy)" ]]; then 
+    docker rm -f reverse_proxy
+    docker-compose -f docker/docker-compose.yml up -d reverse_proxy
+    sleep 5
+fi
+
 echo "Creating self signed SSL certificate complete!"
 
-docker rm -f reverse_proxy
-docker-compose -f docker/docker-compose.yml up -d reverse_proxy
+
+echo "Creating simple local DNS zone..."
 
 current_date=$(date +%Y%m%d)
-echo "Creating simple local DNS zone..."
 zone_name="db.$domain_name"
+
 cat << EOF > bind9/${zone_name}
 \$TTL 300
 @       IN     SOA    ns1.${domain_name}. root.${domain_name}. (
@@ -152,7 +217,11 @@ zone "${domain_name}" IN {
     allow-update { none; };
 };
 EOF
-echo "Creating simple local DNS zone complete!"
 
-docker rm -f bind9
-docker-compose -f docker/docker-compose.yml up -d bind9
+if [[ "$(docker ps -aqf name=bind9)" ]]; then 
+    docker rm -f bind9
+    docker-compose -f docker/docker-compose.yml up -d bind9
+    sleep 5
+fi
+
+echo "Creating simple local DNS zone complete!"
